add + edit vocation without test

+++ app/src/main/java/com/byte4b/judebo/utils/Setting.kt
@@ -94,9 +94,9 @@ class Setting(ctx: Context) {
         //API LEVEL -> (increase const * size)
         val CURRENCY_ICON_SIZE =
             when (Build.VERSION.SDK_INT) {
-                26 -> (1.2 * 24).toInt()//8.0 android
-                27 -> (1.2 * 24).toInt()//8.1
-                28 -> 24 / 2//9
+                26 -> (1.2 * 24).toInt() //8.0 android
+                27 -> (1.2 * 24).toInt() //8.1
+                28 -> (24.0 / 2).toInt() //9
 
                 else -> 24//other versions
             }
